On branch main  Your branch is up to date with 'origin/main'.
 Changes to be committed:
	modified:   automation/export_ppt.py
	modified:   automation/filter.py
	modified:   automation/pipeline.py
	new file:   automation/s3_integration.py
	modified:   dataconfig/data.py
	modified:   dataconfig/db_config.py
	modified:   dataconfig/mapping.py
	modified:   main.py
	modified:   requirements.txt
	modified:   template/query_templates.py
	modified:   template/template_infografis.pptx
	modified:   template/template_infografis_nasional.pptx

+++ automation/pipeline.py
@@ -4,17 +4,18 @@ from automation.export_excel import export_per_provinsi, export_nasional
 from automation.export_ppt import isi_infografis, generate_instansi_values, isi_balok_pendidikan, sanitize_filename
 from pptx import Presentation
 import pandas as pd
-from dataconfig.mapping import mapping_placeholder_per_kategori, label_column_per_kategori, selid_kategori_shared
+from dataconfig.mapping import mapping_placeholder_per_kategori, label_column_per_kategori, selid_kategori_shared, logger
 from io import BytesIO
 '''from google.oauth2.credentials import Credentials
 from google_auth_oauthlib.flow import InstalledAppFlow
 from googleapiclient.discovery import build
 from googleapiclient.http import MediaIoBaseUpload'''
-from template.query_templates import get_instansi_list, get_all_instansi
+from template.query_templates import get_instansi_list, get_all_instansi, get_provinsi_prefix, get_kanreg_prefix, get_all_provinsi
 import os, sys
 import subprocess
 from automation.filter import resolve_filter, normalize_args
 from datetime import datetime
+from automation.s3_integration import upload_file_s3, save_to_db
 
 #TARIK DATA
 def run_pipeline(args):
@@ -27,10 +28,10 @@ def run_pipeline(args):
         mode = 'nasional'
         target_list = [{}]
         nama_table = get_table_name()
-        print("\nNama tabel yang digunakan:", nama_table,"\n")
+        logger.info(f"Nama tabel yang digunakan: {nama_table}")
         data_nasional = get_all_data(conn, nama_table, target_list, mode=mode)
         prs = Presentation('template/template_infografis_nasional.pptx')
-        print("Memproses Data Nasional")
+        logger.info("Memproses Data Nasional")
         kategori_items = list(data_nasional["Nasional"].items())
         for kategori, df in kategori_items:
             if 'count' in df.columns:
@@ -65,21 +66,6 @@ def run_pipeline(args):
                 slide = prs.slides[slide_index]
                 df_export = data_nasional["Nasional"][kategori]
                 isi_balok_pendidikan(df_export, slide, mode)
-            #pie chart
-            '''if kategori == "Kelompok_Generasi":
-                slide_index = selid_kategori_shared.get(kategori)
-                slide = prs.slides[slide_index]
-                df_export = data_nasional["Nasional"][kategori]
-                generate_pie_chart(
-                    df_export,
-                    label_column="kelompok_generasi",
-                    value_column="count",
-                    slide=slide,
-                    left_cm=18.37,
-                    top_cm=3.38,
-                    width_cm=6.65,
-                    height_cm=4.24
-                )'''
             #Isi Data PPT
             if kategori in mapping_placeholder_per_kategori:
                 mapping_placeholder = mapping_placeholder_per_kategori[kategori]
@@ -91,16 +77,16 @@ def run_pipeline(args):
                     '''print(f"Memproses isi_infografis untuk kategori: {kategori}")'''
                     isi_infografis(prs, data_export, mapping_placeholder, mapping_instansi, label_column= label_column)
                 else:
-                    print(f"Data Kosong untuk Nasional, Kategori: {kategori}")
+                    logger.warning(f"Data Kosong untuk Nasional, Kategori: {kategori}")
         #Save PPT
         output_path = "output/ppt/Statistik_Nasional.pptx"
         prs.save(output_path)
-        print(f"PowerPoint Disimpan: {output_path}")
+        logger.info(f"PowerPoint Disimpan: {output_path}")
         #Export ke Excel
         '''if data_nasional and any(data_nasional.values()):
             export_per_provinsi(data_nasional)
         else:
-            print("Tidak ada data yang tersedia untuk diekspor.")'''
+            logger.info("Tidak ada data yang tersedia untuk diekspor.")'''
         #Convert JPG
         jpg_output_dir = f"output/jpg/"
         jpg_output_file = os.path.join(jpg_output_dir, "Statistik_Nasional.jpg")
@@ -112,11 +98,10 @@ def run_pipeline(args):
             output_path
         ]
         subprocess.run(cmd, check=True)
-        print(f"Slide diexport as JPG di {jpg_output_dir}")
+        logger.info(f"Slide diexport as JPG di {jpg_output_dir}")
         '''if os.path.isfile(output_path):
             os.remove(output_path)
-            print(f"File PPT {output_path} dihapus")'''
-
+            logger.info(f"File PPT {output_path} dihapus")'''
         sys.exit()
 
     #MODE INSTANSI
@@ -128,23 +113,26 @@ def run_pipeline(args):
         target_list = get_instansi_list(conn, filter_value)
     elif args.get("provinsi"):
         mode = 'provinsi'
-        target_list = filter_value
+        target_list = get_provinsi_prefix(conn, filter_value)
     elif args.get("kanreg"):
         mode = 'kanreg'
-        target_list = filter_value
+        target_list = get_kanreg_prefix(conn,filter_value)
     elif args.get("all"):
         mode = 'all'
         target_list = get_all_instansi(conn)
+    elif args.get("allprovinsi"):
+        mode = 'allprovinsi'
+        target_list = get_all_provinsi(conn)
     else:
         raise ValueError("Mode harus 'instansi', 'provinsi', atau 'kanreg'")
     nama_table = get_table_name()
-    print("\nNama tabel yang digunakan:", nama_table,"\n")
+    logger.info(f"Nama tabel yang digunakan: {nama_table}")
     data_dict = get_all_data(conn, nama_table, target_list or [], mode=mode)
     # results = []
     for label, kategori_data in data_dict.items():
         prs = Presentation('template/template_infografis.pptx')
-        print(f"Data untuk Instansi: {label}")
-        if mode == "all":
+        logger.info(f"Data siap untuk Instansi: {label}")
+        if mode in ("all","allprovinsi"):
             kategori_items = list(data_dict[label]["kategori"].items())
         else:
             kategori_items = list(kategori_data.items())
@@ -170,17 +158,17 @@ def run_pipeline(args):
                 #Penambahan Ringkasan Kelompok Jabatan       
                 elif 'kelompok_jabatan' in df.columns and 'count' in df.columns:
                     df_ringkasan = ringkasan_kelompok_jabatan(df)
-                    if mode == "all" :
+                    if mode in ("all","allprovinsi") :
                         data_dict[label]["kategori"][f"Ringkasan_{kategori}"] = df_ringkasan
                     else : 
                         data_dict[label][f"Ringkasan_{kategori}"] = df_ringkasan
                 else:
-                    if mode == "all" :
+                    if mode in ("all","allprovinsi") :
                         data_dict[label]["kategori"][kategori] = tambah_persentase(df)
                     else :
                         data_dict[label][kategori] = tambah_persentase(df)
         #Add Ulang Kategori  
-        if mode == "all":         
+        if mode in ("all","allprovinsi"):         
             kategori_items = list(data_dict[label]["kategori"].items())
         else:
             kategori_items = list(kategori_data.items())
@@ -188,7 +176,7 @@ def run_pipeline(args):
             if kategori == "Pendidikan":
                 slide_index = selid_kategori_shared.get(kategori)
                 slide = prs.slides[slide_index]
-                if mode == "all":
+                if mode in ("all","allprovinsi"):
                     df_export = data_dict[label]["kategori"][kategori]
                 else: 
                     df_export = data_dict[label][kategori]
@@ -199,25 +187,26 @@ def run_pipeline(args):
                 label_column = label_column_per_kategori.get(kategori, "label")
                 #Nama Instansi Dinamis                  
                 mapping_instansi = generate_instansi_values(label)
-                if mode == "all":
+                if mode in ("all","allprovinsi"):
                     data_export = data_dict[label]["kategori"][kategori]
                 else:
                     data_export = data_dict[label][kategori]
                 if data_export is not None:
                     isi_infografis(prs, data_export, mapping_placeholder, mapping_instansi, label_column= label_column)
                 else:
-                    print(f"Data Kosong untuk Provinsi: {label}, Kategori: {kategori}")
+                    logger.warning(f"Data Kosong untuk Provinsi: {label}, Kategori: {kategori}")
         #Save PPT
         safe_label = sanitize_filename(label.strip())
         output_path = f"output/ppt/{safe_label}.pptx"
         prs.save(output_path)
-        print(f"PowerPoint Disimpan: {output_path}")
+        logger.info(f"PowerPoint Disimpan: {output_path}")
 
         #Convert JPG
-        if mode == "all":
+        if mode in ("all","allprovinsi"):
             instansi_id = data_dict[label]["instansi_id"]
         periode = datetime.now().strftime("%Y%m")
-        if mode == "all":
+        periode_db = datetime.now().replace(day=1).date()
+        if mode in ("all","allprovinsi"):
             jpg_output_dir = f"output/jpg/{instansi_id}"  
         else:
             jpg_output_dir = f"output/jpg/"
@@ -230,19 +219,28 @@ def run_pipeline(args):
         ]
         subprocess.run(cmd, check=True)
         #Rename JPG
-        if mode == "all":
+        if mode in ("all","allprovinsi"):
             default_jpg = os.path.join(jpg_output_dir, f"{safe_label}.jpg")
             rename_jpg = os.path.join(jpg_output_dir, f"{periode}.jpg")
             if os.path.exists(default_jpg):
                 os.rename(default_jpg, rename_jpg)
-            print(f"Slide diexport as JPG di {rename_jpg}")
+                '''logger.info(f"Slide diexport as JPG di {rename_jpg}")'''
+            #S3
+            local_file = rename_jpg
+            if mode == "all":
+                s3_path = f"/instansi/{instansi_id}/{periode}.jpg"
+            elif mode == "allprovinsi":
+                s3_path = f"/provinsi/{instansi_id}/{periode}.jpg"
+            s3_url = upload_file_s3(local_file, s3_path)
+            logger.info(f"File tersimpan di: {s3_url}")
+            save_to_db(periode_db, instansi_id, s3_path)
         else:
-            print(f"Slide diexport as JPG di {jpg_output_dir}")
-
+            logger.info(f"Slide diexport as JPG di {jpg_output_dir}")
         #Hapus PPT
-        '''if os.path.isfile(output_path):
-            os.remove(output_path)
-            print(f"File PPT {output_path} dihapus")'''
+        if mode in ("all","allprovinsi"):
+            if os.path.isfile(output_path):
+                os.remove(output_path)
+                logger.info(f"File PPT {output_path} dihapus")
 
         #return hasil
         '''results.append({
